requirements.txt
Django == 4.1.7
Pillow
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
PyJWT